201: Analytics empty data - common annotations - Upgrade development version, minor changes to `1.1.0-SNAPSHOT` - Prepare for the next release `1.1.0` - Upgrade `ob-common` version to `1.1.0` - Added new dependency `annotations` from `ob-common` - Refactor of all classes to use the new origin of annotations
Issue: https://github.com/OpenBankingToolkit/openbanking-analytics/issues/201
