+++ forgerock-openbanking-analytics-client/src/main/java/com/forgerock/openbanking/analytics/model/entries/EndpointUsageAggregate.java
@@ -24,8 +24,8 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import com.forgerock.openbanking.analytics.model.kpi.EndpointStatisticKPI;
-import com.forgerock.openbanking.analytics.model.openbanking.OBGroupName;
-import com.forgerock.openbanking.analytics.model.openbanking.OBReference;
+import com.forgerock.openbanking.api.annotations.OBGroupName;
+import com.forgerock.openbanking.api.annotations.OBReference;
 import com.forgerock.openbanking.model.UserContext;
 import com.forgerock.openbanking.serialiser.IsoDateTimeDeserializer;
 import com.forgerock.openbanking.serialiser.IsoDateTimeSerializer;
