+++ forgerock-openbanking-analytics-client/src/main/java/com/forgerock/openbanking/analytics/model/entries/EndpointUsageEntry.java
@@ -22,8 +22,8 @@ package com.forgerock.openbanking.analytics.model.entries;
 
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
-import com.forgerock.openbanking.analytics.model.openbanking.OBGroupName;
-import com.forgerock.openbanking.analytics.model.openbanking.OBReference;
+import com.forgerock.openbanking.api.annotations.OBGroupName;
+import com.forgerock.openbanking.api.annotations.OBReference;
 import com.forgerock.openbanking.model.UserContext;
 import com.forgerock.openbanking.serialiser.IsoDateTimeDeserializer;
 import com.forgerock.openbanking.serialiser.IsoDateTimeSerializer;
