--- forgerock-openbanking-analytics-client/src/main/java/com/forgerock/openbanking/analytics/model/openbanking/OpenBankingAPI.java
@@ -1,38 +1,0 @@
-/**
- * Copyright 2019 ForgeRock AS.
- *
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-package com.forgerock.openbanking.analytics.model.openbanking;
-
-import org.springframework.web.bind.annotation.RequestMapping;
-
-import java.lang.annotation.*;
-
-@Target({ElementType.METHOD, ElementType.TYPE})
-@Retention(RetentionPolicy.RUNTIME)
-@Documented
-@RequestMapping
-public @interface OpenBankingAPI {
-
-    String obVersion() default "";
-
-    OBReference obReference() default OBReference.NONE;
-
-    OBGroupName obGroupName() default OBGroupName.NONE;
-}
